rl_pricing_project/requirements.txt
# RL Pricing Project - Dependencies

# Core ML & RL
torch>=2.0.0
stable-baselines3>=2.0.0
gymnasium>=0.29.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Database
mysql-connector-python>=8.1.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# HuggingFace Hub (NOUVEAU - pour les modèles pré-entraînés)
huggingface-hub>=0.16.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.9.0
flake8>=6.1.0
pytest>=7.4.0

# Utils
tqdm>=4.66.0
tensorboard>=2.14.0